Experimental: shared input files, but separate submit file per function
For each function @job we now generate a separate submit file; that means
it's easier to tweak parameters for just those jobs. But if you create
many instances of that function, all the args go into a single file. This
potentially generates a large file, but it means (a) large objects which
are passed to all invocations are shared, and (b) a much smaller number
of input files.