Revert 59f4920
I now realise that
    error = $(JOB).err
    output = $(JOB).out
doesn't work if you want to use the output of one node as the input to another - you need
to have an explicit filename for the output file. If you are running multiple DAGs
(using splice, subdag external, or multiple DAGs on the condor_submit_dag command line)
then name munging takes place.  This means $(JOB) is unique [good] - but we don't know what
it will be at the time we write the DAG [bad] :-(

This also fixes the test suite