Use less sophisticated argument dependencies.
Now we only automatically add input and parent dependencies for top-level
arguments which are themselves jobs - no longer using pickle to search
recursively.  This is because the pickling was very expensive when given
large argument objects, and could add hours of delay to a DAG with tens of
thousands of nodes.